+++ trl/experimental/prm/prm_trainer.py
@@ -35,9 +35,9 @@ from transformers import (
 from transformers.trainer_utils import EvalPrediction
 from transformers.utils import is_peft_available
 
-from ...models import prepare_peft_model
 from ...trainer.base_trainer import BaseTrainer
 from ...trainer.utils import disable_dropout_in_model
+from ..utils import prepare_peft_model
 from .prm_config import PRMConfig
